Fix reverseY and noData for RGB rendering
I've found data that doesn't render right as RGB, and tracked it down to
the lack of reverseY correction. Resolved by moving the correction below
the data fetch, since it doesn't matter if it's RGB or not.

Also fixed the noData value on plotty for the RGB plot since it was
missing.